Add *.iso to .gitignore  |  Fix test failure message  |  Add -enable-kvm flag to qemu for tests

+++ tests/commit-tests-x86_32.sh
@@ -10,7 +10,9 @@ BIN_DIR=build/run/iso/boot
 GRUB_CFG=$BIN_DIR/grub/grub.cfg
 ISO=all_tests_multiboot.iso
 
-rm $GRUB_CFG
+if [ $GRUB_CFG]; then 
+	rm $GRUB_CFG
+fi
 
 make clean
 make 
@@ -23,6 +25,6 @@ done
 
 if [ $(ls $BIN_DIR/*.bin | wc -l) -gt 0   ];then 
 	grub-mkrescue -o  $ISO build/run/iso
-	qemu-system-i386 -cdrom $ISO  -monitor stdio #-enable-kvm
+	qemu-system-i386 -cdrom $ISO  -monitor stdio -enable-kvm
 fi
